backend/requirements.txt
fastapi
uvicorn[standard]
supabase
anthropic
httpx
yfinance
python-dotenv
feedparser
